Correccion de errores

+++ Producto.java
@@ -1,10 +1,10 @@
 public class Producto {
     private int idProducto;
     private String nombre;
-    private double precio;
+    private int precio;
     private String descripcion;
 
-    public Producto(int idProducto, String nombre, double precio, String descripcion) {
+    public Producto(int idProducto, String nombre, int precio, String descripcion) {
         this.idProducto = idProducto;
         this.nombre = nombre;
         this.precio = precio;
@@ -19,7 +19,7 @@ public class Producto {
         return nombre;
     }
 
-    public double getPrecio() {
+    public int getPrecio() {
         return precio;
     }
